Replace React frontend with Thymeleaf Tailwind version
