--- frontend/README.md
@@ -1,5 +1,0 @@
-# CloudLocker Frontend Components
-
-These React + TypeScript components implement a sample UI for the CloudLocker dashboard using Tailwind CSS and Heroicons.
-
-Place these files in a Vite React project with Tailwind configured and import `App.tsx` in your `main.tsx`.
